\name{miscEx}
\alias{gmF8}
\alias{pedF8}
\alias{gmapF8}
\alias{gdatF8}
\alias{pdatF8}
\alias{miscEx}
\docType{data}
\title{
  Genotype data, phenotype data, genetic map and pedigree.
}
\description{
AIL F8 data include the following:

"gmF8": A list with elements inbreeding coefficients "ib", additive genetic matrix "AA", dominance genetic matrix "DD" and other genetic matrices.

"pedF8": Pedigree date.

"gmapF8: Genetic map.

"gdatF8": Genotype data.

"pdatF8": Phenotype data.
}
\usage{
data(miscEx)
}
\keyword{datasets}
