\name{miscEx}
\alias{y}
\alias{x}
\alias{v}
\alias{ped}
\alias{genMap}
\alias{genMapF34}
\alias{gdat}
\alias{miscEx}
\docType{data}
\title{
  Pedigree, genotype data and genetic map.
}
\description{
  Pedigree, genotype data, genetic map and other data that are used in examples.
}
\usage{
data(miscEx)
}
\keyword{datasets}
