\name{miscEx}
\alias{gmF8}
\alias{pedF8}
\alias{gmapF8}
\alias{gdatF8}
\alias{pdatF8}
\alias{miscEx}
\docType{data}
\title{
  Genotype data, phenotype data, genetic map and pedigree.
}
\description{
AIL F8 data include the following:

"gmF8": a list with elements inbreeding coefficients "ib", additive genetic matrix "AA", dominance genetic matrix "DD" and other genetic matrices.

"pedF8": pedigree date.

"gmapF8: genetic map.

"gdatF8": genotype data.

"pdatF8": phenotype data.
}
\usage{
data(miscEx)
}
\keyword{datasets}
